refactor: dto 패키지 세부 분류
