+++ src/main/java/subway/application/SectionService.java
@@ -12,9 +12,9 @@ import subway.dao.entity.SectionEntity;
 import subway.domain.Section;
 import subway.domain.Sections;
 import subway.domain.Station;
-import subway.dto.SectionCreateRequest;
-import subway.dto.SectionDeleteRequest;
-import subway.dto.SectionResponse;
+import subway.dto.section.SectionCreateRequest;
+import subway.dto.section.SectionDeleteRequest;
+import subway.dto.section.SectionResponse;
 import subway.exception.IllegalDistanceException;
 import subway.exception.IllegalSectionException;
 import subway.exception.StationNotFoundException;
