+++ src/main/java/subway/dto/section/SectionResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.section;
 
 import subway.domain.Section;
