+++ src/main/java/subway/ui/SectionController.java
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import subway.application.SectionService;
-import subway.dto.SectionCreateRequest;
-import subway.dto.SectionDeleteRequest;
-import subway.dto.SectionResponse;
+import subway.dto.section.SectionCreateRequest;
+import subway.dto.section.SectionDeleteRequest;
+import subway.dto.section.SectionResponse;
 
 @RestController
 @RequestMapping("/sections")
