+++ src/test/java/subway/integration/SectionControllerIntegrationTest.java
@@ -18,8 +18,8 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
-import subway.dto.SectionCreateRequest;
-import subway.dto.SectionDeleteRequest;
+import subway.dto.section.SectionCreateRequest;
+import subway.dto.section.SectionDeleteRequest;
 
 @SpringBootTest
 @Transactional
