+++ src/test/java/subway/ui/SectionControllerTest.java
@@ -16,8 +16,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import subway.application.SectionService;
-import subway.dto.SectionCreateRequest;
-import subway.dto.SectionDeleteRequest;
+import subway.dto.section.SectionCreateRequest;
+import subway.dto.section.SectionDeleteRequest;
 
 @WebMvcTest(SectionController.class)
 class SectionControllerTest {
